+++ main.cs
@@ -157,10 +157,6 @@ public class Main : BasePlugin
     public static Dictionary<byte, int> GuesserGuessed = [];
     public static Dictionary<(byte, byte), bool> isDraw = [];
     public static Dictionary<(byte, byte), bool> isRevealed = [];
-    public static Dictionary<byte, (PlayerControl, float)> RevolutionistTimer = [];
-    public static Dictionary<byte, long> RevolutionistStart = [];
-    public static Dictionary<byte, long> RevolutionistLastTime = [];
-    public static Dictionary<byte, int> RevolutionistCountdown = [];
     public static Dictionary<byte, byte> SpeedBoostTarget = [];
     public static Dictionary<byte, int> ParaUsedButtonCount = [];
     public static Dictionary<byte, int> MarioVentCount = [];
